+++ how-to/integrate-with-bloomberg-basic/client/src/bootstrapper.ts
@@ -1,0 +1,6 @@
+export async function init() {
+    const providerWindow = fin.Window.getCurrentSync();
+	await providerWindow.once("close-requested", async (event) => {
+		await fin.Platform.getCurrentSync().quit();
+	});
+}
