+++ how-to/integrate-with-bloomberg-basic/client/src/interopbroker.ts
@@ -1,0 +1,16 @@
+export function interopOverride(
+	InteropBroker: OpenFin.Constructor<OpenFin.InteropBroker>,
+	provider?: OpenFin.ChannelProvider,
+	options?: OpenFin.InteropBrokerOptions,
+	...args: unknown[]
+): OpenFin.InteropBroker {
+    class InteropOverride extends InteropBroker {
+        public async handleFiredIntent(intent: OpenFin.Intent) {
+            console.log("Received request for a raised intent:", intent);
+            const targetIdentity = fin.me.identity;
+            await super.setIntentTarget(intent, targetIdentity);
+        }
+    }
+
+    return new InteropOverride();
+}
