+++ how-to/integrate-with-bloomberg-basic/public/platform/provider.html
@@ -11,51 +11,17 @@
 	</head>
 
 	<body class="col fill gap20">
-		<header class="row spread middle">
-			<div class="col">
-				<h1>Integrate with Bloomberg Terminal</h1>
-				<h1 class="tag">Demonstrate integrating with the Bloomberg Terminal.</h1>
-			</div>
-			<div class="row middle gap20">
-				<image src="/common/images/icon-blue.png" alt="OpenFin" height="40px"></image>
-			</div>
-		</header>
+		
 		<main class="col fill gap20 scroll-vertical">
-			<p>
-				Under normal conditions you would keep this window hidden by setting autoShow to false in the
-				manifest.
-			</p>
-			<div class="row gap20">
-				<pre id="logOutput" class="fill scroll-vertical" style="min-height: 130px; max-height: 110px;"></pre>
-			</div>
-			<hr />
-
-			<div class="col gap5">
-				<h3>
-					Fire Intents:
-				</h3>
-				<fieldset class="row">
-					<label for="intentName">Intent Name:</label>
-					<select id="intentName">
-						<option value="ViewChart">View Chart</option>
-						<option value="ViewContact">View Contact</option>
-						<option value="ViewInstrument">View Instrument</option>
-						<option value="ViewQuote">View Quote</option>
-					</select>
-					<label for="intentValue">Value to pass for the Intent:</label>
-					<input type="text" id="intentValue"/>
-					<button type="button" id="btnQuery">Go</button>
-					<br>
-				</fieldset>								
-			</div>
-		</main>
-		<footer>
 			<div class="row spread">
-				<button id="btnClear">Clear Log</button>
+				
 				<div class="row gap10">
 					<button id="btnConnect">Connect To BBG Terminal</button>
 				</div>
 			</div>
-		</footer>
+
+			
+		</main>
+		
 	</body>
 </html>
